requirements_mask_anomaly.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appdirs=1.4.4=pypi_0
black=21.4b2=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7f98852_4
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
cityscapesscripts=2.2.2=pypi_0
click=8.1.7=pypi_0
cloudpickle=3.0.0=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.1.1=pypi_0
cudatoolkit=11.1.1=ha002fc5_10
cycler=0.12.1=pypi_0
cython=3.0.8=pypi_0
detectron2=0.6+cu111=pypi_0
easydict=1.11=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=pypi_0
fonttools=4.47.2=pypi_0
freetype=2.10.4=h0708190_1
fsspec=2023.12.2=pypi_0
future=0.18.3=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=2.27.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
grpcio=1.60.0=pypi_0
h5py=3.10.0=pypi_0
huggingface-hub=0.20.3=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.6=pypi_0
imageio=2.33.1=pypi_0
importlib-metadata=7.0.1=pypi_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
iopath=0.1.9=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jpeg=9b=h024ee3a_2
kiwisolver=1.4.5=pypi_0
kornia=0.5.2=pypi_0
lame=3.100=h7f98852_1001
lazy-loader=0.3=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.1.0=h2733197_1
libuv=1.43.0=h7f98852_0
lovasz-losses=0.1=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
markdown=3.5.2=pypi_0
markupsafe=2.1.4=pypi_0
matplotlib=3.7.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h95df7f1_0
mkl_fft=1.3.1=py38h8666266_1
mkl_random=1.2.2=py38h1abd341_0
mpmath=1.3.0=pypi_0
multiscaledeformableattention=1.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nettle=3.6=he412f7d_0
networkx=3.1=pypi_0
ninja=1.11.0=h924138e_0
numpy=1.24.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
olefile=0.47=pyhd8ed1ab_0
omegaconf=2.3.0=pypi_0
ood-metrics=1.1.1=pypi_0
opencv-python=4.9.0.80=pypi_0
openh264=2.1.1=h780b84a_0
openssl=3.0.13=h7f8727e_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
pathspec=0.12.1=pypi_0
pillow=9.5.0=pypi_0
pip=23.3.1=py38h06a4308_0
portalocker=2.8.2=pypi_0
protobuf=4.25.2=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycocotools=2.0.7=pypi_0
pydot=2.0.0=pypi_0
pyparsing=3.1.1=pypi_0
pyquaternion=0.9.9=pypi_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=1.8.0=py3.8_cuda11.1_cudnn8.0.5_0
pytz=2024.1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=pypi_0
readline=8.2=h5eee18b_0
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.13.2=pypi_0
setuptools=68.2.2=py38h06a4308_0
shapely=2.0.2=pypi_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h5eee18b_0
submitit=1.5.1=pypi_0
sympy=1.12=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.14.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.7.10=pypi_0
timm=0.9.12=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pypi_0
torch=2.2.0=pypi_0
torchaudio=0.8.0=py38
torchvision=0.9.0=py38_cu111
tqdm=4.66.1=pypi_0
triton=2.2.0=pypi_0
typing=3.7.4.3=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023.4=pypi_0
urllib3=2.1.0=pypi_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py38h06a4308_0
xz=5.4.5=h5eee18b_0
yacs=0.1.8=pypi_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=ha95c52a_0
